# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoftTrack",
    "description": "An open-source, self-hostable issue tracker inspired by Linear.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/teams/{team_id}/views": {
      "get": {
        "tags": [
          "views"
        ],
        "summary": "List Views",
        "description": "Every view this person can see on the team, plus which one is default.\n\nNot paginated. A team that has more saved views than fit in one response\nhas a different problem, and the sidebar renders all of them at once.",
        "operationId": "list_views_teams__team_id__views_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViews"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "views"
        ],
        "summary": "Create View",
        "operationId": "create_view_teams__team_id__views_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/views/{view_id}": {
      "patch": {
        "tags": [
          "views"
        ],
        "summary": "Update View",
        "operationId": "update_view_views__view_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "View Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedViewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViewRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "views"
        ],
        "summary": "Delete View",
        "operationId": "delete_view_views__view_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "view_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "View Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/default-view": {
      "put": {
        "tags": [
          "views"
        ],
        "summary": "Set Team Default View",
        "description": "Where everyone on the team lands. Team admins only, shared views only.\n\nReturns the whole list rather than the view, because setting a default\nalso unsets the previous one and the sidebar has to redraw both.",
        "operationId": "set_team_default_view_teams__team_id__default_view_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultViewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViews"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/{team_id}/default-view/me": {
      "put": {
        "tags": [
          "views"
        ],
        "summary": "Set My Default View",
        "description": "Where *you* land on this team, overriding the team's default.\n\nClearing it falls back to the team's rather than to all issues: no\noverride means no preference, not a preference for nothing.",
        "operationId": "set_my_default_view_teams__team_id__default_view_me_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultViewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedViews"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DefaultViewUpdate": {
        "properties": {
          "view_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "View Id"
          }
        },
        "type": "object",
        "title": "DefaultViewUpdate",
        "description": "Null clears the default rather than setting one."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssuePriority": {
        "type": "string",
        "enum": [
          "no_priority",
          "urgent",
          "high",
          "medium",
          "low"
        ],
        "title": "IssuePriority"
      },
...
      "IssueStatus": {
        "type": "string",
        "enum": [
          "backlog",
          "todo",
          "in_progress",
          "in_review",
          "done",
          "cancelled"
        ],
        "title": "IssueStatus"
      },
...
      "SavedViewCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Name"
          },
          "is_shared": {
            "type": "boolean",
            "title": "Is Shared",
            "default": false
          },
          "filters": {
            "$ref": "#/components/schemas/ViewFilters",
            "default": {
              "unassigned": false
            }
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SavedViewCreate"
      },
      "SavedViewRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "is_shared": {
            "type": "boolean",
            "title": "Is Shared"
          },
          "filters": {
            "$ref": "#/components/schemas/ViewFilters"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "name",
          "owner",
          "is_shared",
          "filters",
          "created_at",
          "updated_at"
        ],
        "title": "SavedViewRead"
      },
      "SavedViewUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "is_shared": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Shared"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViewFilters"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "SavedViewUpdate",
        "description": "Every field optional: renaming, re-filtering and sharing are separate\ngestures in the UI and each sends only what it changed."
      },
      "SavedViews": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedViewRead"
            },
            "type": "array",
            "title": "Items"
          },
          "team_default_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Default Id"
          },
          "my_default_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "My Default Id"
          },
          "effective_default_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Default Id"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SavedViews",
        "description": "The whole sidebar in one response.\n\nThe defaults come back with the list rather than from their own endpoint\nbecause the sidebar cannot render a row without knowing whether it is\nstarred, and a second request would mean a frame where none of them are."
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "avatar_color": {
            "type": "string",
            "title": "Avatar Color"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "username",
          "full_name",
          "avatar_color",
          "is_active"
        ],
        "title": "UserPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ViewFilters": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssueStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IssuePriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "unassigned": {
            "type": "boolean",
            "title": "Unassigned",
            "default": false
          },
          "label_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "cycle_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cycle Id"
          }
        },
        "type": "object",
        "title": "ViewFilters",
        "description": "What a view narrows the issue list to.\n\nNested rather than flattened into the view models so that the same shape\nis what the board sends, what a view stores, and what the frontend keeps\nin the URL -- one type to keep in step instead of three.\n\nNull everywhere means \"all issues\", which is what an empty view is."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}